+++ impacted-test-engine/src/main/java/com/teamscale/test_impacted/engine/ImpactedTestEngineConfiguration.java
@@ -1,28 +1,31 @@
 package com.teamscale.test_impacted.engine;
 
-import com.teamscale.test_impacted.engine.executor.ITestExecutor;
+import com.teamscale.test_impacted.engine.executor.ITestSorter;
+import com.teamscale.test_impacted.engine.executor.TeamscaleAgentNotifier;
 import org.junit.platform.engine.TestEngine;
-
-import java.io.File;
 
 /** Container for a configuration used by the {@link ImpactedTestEngine} */
 public class ImpactedTestEngineConfiguration {
 
 	/** The directory to write testwise coverage and available tests to. */
-	final File reportDirectory;
+	final TestDataWriter testDataWriter;
 
 	/** The test engine registry used to determine the {@link TestEngine}s to use. */
 	final TestEngineRegistry testEngineRegistry;
 
-	/** The {@link ITestExecutor} to use for execution of tests. */
-	final ITestExecutor testExecutor;
+	/** The {@link ITestSorter} to use for execution of tests. */
+	final ITestSorter testSorter;
+
+	/** An API to signal test start and end to the agent. */
+	final TeamscaleAgentNotifier teamscaleAgentNotifier;
 
 	public ImpactedTestEngineConfiguration(
-			File reportDirectory,
+			TestDataWriter testDataWriter,
 			TestEngineRegistry testEngineRegistry,
-			ITestExecutor testExecutor) {
-		this.reportDirectory = reportDirectory;
+			ITestSorter testSorter, TeamscaleAgentNotifier teamscaleAgentNotifier ) {
+		this.testDataWriter = testDataWriter;
 		this.testEngineRegistry = testEngineRegistry;
-		this.testExecutor = testExecutor;
+		this.testSorter = testSorter;
+		this.teamscaleAgentNotifier = teamscaleAgentNotifier;
 	}
 }