+++ impacted-test-engine/src/main/java/com/teamscale/test_impacted/test_descriptor/TestDescriptorResolverRegistry.java
@@ -1,13 +1,12 @@
 package com.teamscale.test_impacted.test_descriptor;
 
-import org.junit.platform.commons.logging.Logger;
-import org.junit.platform.commons.logging.LoggerFactory;
+import com.teamscale.test_impacted.commons.LoggerUtils;
 import org.junit.platform.commons.util.ClassLoaderUtils;
-import org.junit.platform.engine.TestEngine;
 
 import java.util.HashMap;
 import java.util.Map;
 import java.util.ServiceLoader;
+import java.util.logging.Logger;
 
 /**
  * Registry containing the default and custom {@link ITestDescriptorResolver}s discovered by the java
@@ -15,7 +14,7 @@ import java.util.ServiceLoader;
  */
 public class TestDescriptorResolverRegistry {
 
-	private static final Logger LOGGER = LoggerFactory.getLogger(TestDescriptorResolverRegistry.class);
+	private static final Logger LOGGER = LoggerUtils.getLogger(TestDescriptorResolverRegistry.class);
 
 	private static final Map<String, ITestDescriptorResolver> TEST_DESCRIPTOR_RESOLVER_BY_ENGINE_ID = new HashMap<>();
 
@@ -36,11 +35,11 @@ public class TestDescriptorResolverRegistry {
 	}
 
 	/** Returns the test descriptor resolver or null if none exists for the test engine. */
-	public static ITestDescriptorResolver getTestDescriptorResolver(TestEngine testEngine) {
-		String testEngineId = testEngine.getId();
+	public static ITestDescriptorResolver getTestDescriptorResolver(String testEngineId) {
 		if (!TEST_DESCRIPTOR_RESOLVER_BY_ENGINE_ID.containsKey(testEngineId)) {
-			LOGGER.warn(() -> testEngineId + " is not officially supported! You can add support by " +
-					"implementing the ITestDescriptorResolver interface and making the implementation via the Java Service Loader mechanism!");
+			LOGGER.warning(() -> testEngineId + " is not officially supported! You can add support by " +
+					"implementing the ITestDescriptorResolver interface and making the implementation " +
+					"discoverable via the Java Service Loader mechanism!");
 			return TEST_DESCRIPTOR_RESOLVER_BY_ENGINE_ID.get("junit-jupiter");
 		}
 		return TEST_DESCRIPTOR_RESOLVER_BY_ENGINE_ID.get(testEngineId);